Move build/buildrun param validation
Move `BuildParameters` and `BuildRunParameters` into the `validate` package.

Move respecitve test cases into `validate` package.

Introduce Ginkgo suite for `validate` package.

Remove `Validate` prefix of functions due to package name (static code
scan finding).

Organize imports to have Kubernetes imports separately.
